refactor(content-creator): S6c.2 rework — integrate daily-7 เข้า /new + ปิด loop + preview robust
browser-truth (ฟีม) เจอ 3 design gap ที่ test เขียว/curl ไม่จับ (verify-real-path ระดับ product flow):
1. /daily7 เป็นเกาะแยก ไม่เชื่อม navigation → รวมเข้า /new (template dropdown เลือก daily-7 →
   render Daily7Authoring inline เหมือน finance) ; ลบเกาะ /daily7 + /daily7-playground
2. preview ไม่ขึ้น — เดิมส่ง payload ผ่าน GET URL+base64+encodeURIComponent (เปราะ length/encoding)
   → เปลี่ยนเป็น POST+body → blob → objectURL (robust) ; preview route GET→POST
3. finalize → PENDING ค้าง ไม่โผล่คิว approve (ไม่มีใคร generate) → finalize endpoint generate ต่อ
   (ภาพ composition + caption) → GENERATED → เด้งไป /content-creator คิว approve (ปิด loop เหมือน finance)

generate บังคับ CTA url (ทุกโพสต์ต้องมี link) — ตั้งใน Settings ก่อน ; ไม่ตั้ง → FAILED + error ชัด.
recovery lifecycle (lib/daily7-session) + tests คงเดิม.

verify: tsc 0 · vitest 230 · next build · **ยิง full loop จริงผ่าน dev**: /new templates มี daily-7,
preview POST→PNG, create→READY, finalize→generate→GENERATED, posts API เห็น post (image+caption),
ดูภาพจริง (การ์ด 7 วัน + วันที่ derive ถูก + ไทยคม) ; client UI = ฝากฟีม browser-truth
