+++ app/content-creator/api/daily/draft/[id]/finalize/route.ts
@@ -1,14 +1,16 @@
 /**
- * POST /content-creator/api/daily/draft/:id/finalize — snapshot → สร้าง contentPost (PENDING) [S6c]
+ * POST /content-creator/api/daily/draft/:id/finalize — snapshot → สร้าง contentPost → generate [S6c.2]
  * body: { finalizeKey, expectedRevision, backgroundId }
- * validate FinalInput strict (7 วัน canonical) + backgroundId อยู่ใน manifest ; atomic กัน double-finalize
- * → คืน contentPostId (ฟีมไปต่อที่ gen ภาพ/approve queue)
+ * 1) finalize: validate FinalInput strict + backgroundId ใน manifest → สร้าง contentPost (PENDING, atomic)
+ * 2) generate: gen ภาพ composition + caption → GENERATED → โผล่คิว approve (ปิด loop เหมือน finance)
+ *    (generate บังคับ CTA url — ไม่ตั้ง → FAILED + error ชัด). replay → generate SKIP ถ้า gen ไปแล้ว
  */
 import { NextResponse } from "next/server";
 import { z } from "zod";
 import { getContentDb } from "@/content-creator/db/client";
 import { isContentCreatorEnabled } from "@/content-creator/lib/enabled";
 import { finalizeDaily7Draft, draftErrorStatus } from "@/content-creator/daily7-service";
+import { generate } from "@/content-creator/engine";
 
 export const runtime = "nodejs";
 export const dynamic = "force-dynamic";
@@ -28,11 +30,21 @@ export async function POST(request: Request, { params }: { params: Promise<{ id:
   } catch {
     return NextResponse.json({ ok: false, error: "invalid body (ต้องมี finalizeKey + expectedRevision + backgroundId)" }, { status: 400 });
   }
+
+  const db = getContentDb();
+  let contentPostId: string;
   try {
-    const res = finalizeDaily7Draft(getContentDb(), id, body.finalizeKey, body.expectedRevision, body.backgroundId);
-    return NextResponse.json({ ok: true, contentPostId: res.contentPostId, idempotent: res.replay }, { status: 200 });
+    contentPostId = finalizeDaily7Draft(db, id, body.finalizeKey, body.expectedRevision, body.backgroundId).contentPostId;
   } catch (e) {
     const { status, error } = draftErrorStatus(e);
     return NextResponse.json({ ok: false, error }, { status });
   }
+
+  // gen ต่อ (sync) → ปิด loop ถึงคิว approve. SKIPPED = gen ไปแล้ว (replay) → ถือว่าสำเร็จ
+  const gen = await generate(db, contentPostId);
+  const generated = gen.ok || gen.status === "SKIPPED";
+  return NextResponse.json(
+    { ok: generated, contentPostId, status: gen.status, caption: gen.caption, error: gen.error },
+    { status: generated ? 200 : 502 }, // 502 = gen ล้ม (เช่น ยังไม่ตั้ง CTA / Gemini) — post เป็น FAILED
+  );
 }
